Extract graph and segmentation helpers
