+++ package.xml
@@ -11,8 +11,8 @@
   <depend>rclcpp</depend>
   <depend>nav_msgs</depend>
   <depend>std_msgs</depend>
-  <depend>opencv4</depend>
-  <depend>yaml-cpp</depend>
+  <build_depend>opencv4</build_depend>
+  <build_depend>yaml-cpp</build_depend>
 
   <exec_depend>opencv4</exec_depend>
   <exec_depend>yaml-cpp</exec_depend>
